slLIN: sllin.h: fixes checksum type

+++ examples/device/sllin/inc/sllin.h
@@ -1,7 +1,7 @@
 /*
  * The MIT License (MIT)
  *
- * Copyright (c) 2021 Jean Gressmann <jean@0x42.de>
+ * Copyright (c) 2021-2022 Jean Gressmann <jean@0x42.de>
  *
  * Permission is hereby granted, free of charge, to any person obtaining a copy
  * of this software and associated documentation files (the "Software"), to deal
@@ -73,7 +73,7 @@ static inline uint_least16_t sllin_crc_update(uint_least16_t crc, uint8_t const 
 
 static inline uint8_t sllin_crc_finalize(uint_least16_t crc)
 {
-	unsigned factor = crc / 256;
+	uint_least16_t factor = crc / 256;
 	crc -= factor * 255;
 
 	return (~crc) & 0xff;
